Feat: 식재료 차감 기능 추가

+++ back/src/main/java/refrigerator/back/ingredient/adapter/in/dto/IngredientRegisterRequestDTO.java
@@ -12,7 +12,7 @@ public class IngredientRegisterRequestDTO {
     // 검증 필요 @Validation
     private String name;                    // 식재료명
     private LocalDate expirationDate;       // 유통기한
-    private Integer capacity;               // 용량
+    private Double capacity;               // 용량
     private String capacityUnit;            // 용량 단위
     private String storageMethod;           // 보관 방식
     private Integer imageId;                // 사진 (수정 가능성)
