+++ back/src/main/java/refrigerator/back/ingredient/adapter/in/dto/IngredientUpdateRequestDTO.java
@@ -11,6 +11,6 @@ import java.time.LocalDate;
 public class IngredientUpdateRequestDTO {
     // 검증 필요 @Validation
     private LocalDate ExpirationDate;       // 유통기한
-    private Integer capacity;               // 용량
+    private Double capacity;               // 용량
     private String storageMethod;           // 보관방법
 }
